caprevoke: Introduce VM_FAULT_NOPMAP
This will let the kernel go through all the gyrations to get a virtual page
into physical memory, but won't expose it (or any others that prefaulting might
otherwise expose) to the pmap / user process.

This is going to be used for a cheap hack to avoid revoking across address
spaces; it probably should not go upstream?